module_29_testing/hw/tests/requirements.txt
amqp==5.3.1
async-timeout==5.0.1
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
exceptiongroup==1.2.2
factory_boy==3.3.3
Faker==37.8.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
kombu==5.4.2
MarkupSafe==3.0.2
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
postgres==4.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
psycopg2-pool==1.2
PyJWT==2.9.0
pytest==8.3.5
python-dateutil==2.9.0.post0
redis==5.3.0b5
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.37
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
WTForms==3.2.1
